retrain all models

+++ PredictModel.py
@@ -9,9 +9,13 @@ from torch.utils.data import DataLoader, TensorDataset
 from torchvision.models.googlenet import GoogLeNet, model_urls
 from torchvision.models.mobilenet import MobileNetV2
 from torchvision.models.mobilenet import mobilenet_v2
+from sklearn.metrics import mean_absolute_error
 import numpy as np
+import os
+from sklearn.metrics import r2_score
 
 
+# dfs遍历所有的碱基组合,并将其编码为64维的向量
 def dfs(current, alphabet, char_map):
     if len(current) == 3:
         index = len(char_map)
@@ -74,8 +78,8 @@ def read_data(path):
     x_encoded = encoding(x, encoder)
     y = torch.tensor(y, dtype=torch.float32).unsqueeze(1)
 
-    # 划分训练集和测试集
-    x_train, x_test, y_train, y_test = train_test_split(x_encoded, y, test_size=0.2, random_state=41)
+    # 划分训练集和测试集 当random_state设置为一个固定的整数值时，每次运行train_test_split函数时，数据的分割结果将保持一致。
+    x_train, x_test, y_train, y_test = train_test_split(x_encoded, y, test_size=0.2, random_state=43)
 
     x_train = x_train.to(device)
     x_test = x_test.to(device)
@@ -85,9 +89,10 @@ def read_data(path):
     return x_train, y_train, x_test, y_test, l
 
 
-class CustomGoogLeNet(GoogLeNet):
+# WrapperGoogleNet
+class WrapperGoogleNet(GoogLeNet):
     def __init__(self, *args, pretrained=False, **kwargs):
-        super(CustomGoogLeNet, self).__init__(*args, **kwargs)
+        super(WrapperGoogleNet, self).__init__(*args, **kwargs)
         if pretrained:
             # 下面的这部分代码在 pretrained=True 时加载预训练权重，并去掉了与辅助分类器（auxiliary classifiers）相关的权重
             state_dict = model_zoo.load_url(model_urls['googlenet'], progress=True)
@@ -103,10 +108,10 @@ class CustomGoogLeNet(GoogLeNet):
         return x
 
 
-class RSBGoogLeNet(nn.Module):
+class GoogleNet(nn.Module):
     def __init__(self, num_classes, input_channels):
-        super(RSBGoogLeNet, self).__init__()
-        self.googlenet = CustomGoogLeNet(pretrained=True, aux_logits=False)
+        super(GoogleNet, self).__init__()
+        self.googlenet = WrapperGoogleNet(pretrained=True, aux_logits=False)
         # 更改输入通道数
         self.googlenet.conv1 = nn.Conv2d(input_channels, 64, kernel_size=(7, 7), stride=(2, 2), padding=(3, 3))
         self.googlenet.fc = nn.Linear(1024, num_classes)
@@ -115,18 +120,18 @@ class RSBGoogLeNet(nn.Module):
         return self.googlenet(x)
 
 
-class CustomMobileNetV2(MobileNetV2):
+class WrapperMobileNet(MobileNetV2):
     def __init__(self, *args, pretrained=False, **kwargs):
-        super(CustomMobileNetV2, self).__init__(*args, **kwargs)
+        super(WrapperMobileNet, self).__init__(*args, **kwargs)
         if pretrained:
             pretrained_model = mobilenet_v2(pretrained=True)
             self.load_state_dict(pretrained_model.state_dict(), strict=False)
 
 
-class RSBMobileNetV2(nn.Module):
+class MobileNet(nn.Module):
     def __init__(self, num_classes, input_channels):
-        super(RSBMobileNetV2, self).__init__()
-        self.mobilenet = CustomMobileNetV2(pretrained=True)
+        super(MobileNet, self).__init__()
+        self.mobilenet = WrapperMobileNet(pretrained=True)
         # 更改输入通道数
         self.mobilenet.features[0][0] = nn.Conv2d(input_channels, 32, kernel_size=(3, 3), stride=(2, 2), padding=(1, 1),
                                                   bias=False)
@@ -181,20 +186,21 @@ def get_predictions(model, dataset, device):
 
 if __name__ == "__main__":
     encoder = 'triplet'
-    data_names = ['promoter', 'rbs']
-    data_name = data_names[1]
+    data_names = ['rbs', 'promoter', 'rbs1', 'promoter1']
+    data_name = data_names[3]
     data_path = r'./data/' + data_name + '-data.csv'
-    model_names = ['RSBGoogLeNet', 'RSBMobileNetV2']
-    modelname = model_names[1]
+    model_names = ['GoogleNet', 'MobileNet']
+    model_name = model_names[1]
     learning_rate = 0.001
     batch_size = 64
-    num_epochs = 80
+    num_epochs = 70
     if encoder == 'one-hot':
         height = 4
     elif encoder == 'triplet':
         height = 64
     num_classes = 1  # 输出类别数
     input_channels = 1  # 输入通道数 一般图像为3，这里由于是序列，所以为1
+    model_path = 'model/' + data_name + '-' + model_name + '_' + encoder + '.pth'
 
     device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
     print('current device is:{}'.format(device))
@@ -210,38 +216,47 @@ if __name__ == "__main__":
     test_loader = DataLoader(test_dataset, batch_size=batch_size, shuffle=False)
 
     # 初始化模型、损失函数、优化器
-    if modelname == 'RSBGoogLeNet':
-        model = RSBGoogLeNet(num_classes, input_channels).to(device)
-    elif modelname == 'RSBMobileNetV2':
-        model = RSBMobileNetV2(num_classes, input_channels).to(device)
+    if model_name == 'GoogleNet':
+        model = GoogleNet(num_classes, input_channels).to(device)
+    elif model_name == 'MobileNet':
+        model = MobileNet(num_classes, input_channels).to(device)
     criterion = nn.MSELoss()
     optimizer = optim.Adam(model.parameters(), lr=learning_rate)
 
-    train_losses = []
-    test_losses = []
-    # 训练模型
-    for epoch in range(num_epochs):
-        train_loss = train(model, train_loader, criterion, optimizer, device)
-        test_loss = evaluate(model, test_loader, criterion, device)
-        train_losses.append(train_loss)
-        test_losses.append(test_loss)
-        if (epoch + 1) % 10 == 0:
-            print(f"Epoch {epoch + 1}/{num_epochs}, Train Loss: {train_loss:.4f}, Test Loss: {test_loss:.4f}")
+    if os.path.exists(model_path):  # 如果模型已存在，加载已有模型
+        print("Loading pre-trained model...")
+        model.load_state_dict(torch.load(model_path))
+        model.eval()
+    else:
+        train_losses = []
+        test_losses = []
+        # 训练模型
+        for epoch in range(num_epochs):
+            train_loss = train(model, train_loader, criterion, optimizer, device)
+            test_loss = evaluate(model, test_loader, criterion, device)
+            train_losses.append(train_loss)
+            test_losses.append(test_loss)
+            if (epoch + 1) % 10 == 0:
+                print(f"Epoch {epoch + 1}/{num_epochs}, Train Loss: {train_loss:.4f}, Test Loss: {test_loss:.4f}")
 
-    plt.plot(train_losses, label='Train Loss')
-    plt.plot(test_losses, label='Test Loss')
-    plt.xlabel('Epoch')
-    plt.ylabel('Loss')
-    plt.legend()
-    pic_name1 = 'pic/' + data_name + '-' + modelname + '_' + encoder + '_loss' + '.png'
-    plt.savefig(pic_name1)
+        plt.plot(train_losses, label='Train Loss')
+        plt.plot(test_losses, label='Test Loss')
+        plt.xlabel('Epoch')
+        plt.ylabel('Loss')
+        plt.legend()
+        pic_name1 = 'pic/' + data_name + '-' + model_name + '_' + encoder + '_loss' + '.png'
+        plt.savefig(pic_name1)
+        torch.save(model.state_dict(), model_path)
 
+    model.eval()  # 切换到评估模式
     train_preds = get_predictions(model, train_dataset, device)
     test_preds = get_predictions(model, test_dataset, device)
+
     pearson_corr = np.corrcoef(y_test.cpu().numpy().flatten(), test_preds.flatten())[0, 1]
+    mae = mean_absolute_error(y_test.cpu().numpy().flatten(), test_preds.flatten())
+    r2 = r2_score(y_test.cpu().numpy().flatten(), test_preds.flatten())
 
     fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(12, 5))
-
     ax1.scatter(y_train.cpu().numpy(), train_preds, alpha=0.5)
     ax1.set_xlabel('True Values')
     ax1.set_ylabel('Predictions')
@@ -256,18 +271,18 @@ if __name__ == "__main__":
     ax2.set_aspect('equal', 'box')  # 添加这一行以使横纵坐标尺寸相同
     ax2.axis('square')  # 添加这一行以使子图呈正方形
 
-    pic_name2 = 'pic/' + data_name + '-' + modelname + '_' + encoder + '_scatter' + '.png'
+    pic_name2 = 'pic/' + data_name + '-' + model_name + '_' + encoder + '_scatter' + '.png'
     plt.savefig(pic_name2)
-
-    model_name = 'model/' + data_name + modelname + '_' + encoder + '.pth'
-    torch.save(model.state_dict(), model_name)
+    # print(model)
 
     with open("log.txt", "a") as log_file:
         log_file.write(f"DataSet: {data_name}\n")
-        log_file.write(f"Model: {modelname}\n")
+        log_file.write(f"Model: {model_name}\n")
         log_file.write(f"Encoder: {encoder}\n")
         log_file.write(f"Num epochs: {num_epochs}\n")
         log_file.write(f"Learning rate: {learning_rate}\n")
         log_file.write(f"Batch size: {batch_size}\n")
+        log_file.write(f"Mean Absolute Error: {mae:.4f}\n")
         log_file.write(f"Pearson Correlation Coefficient: {pearson_corr:.4f}\n")
+        log_file.write(f"R^2 Score: {r2:.4f}\n")  # 添加 R^2 值
         log_file.write("\n")
